# internal/docs/spec/swagger.yaml
basePath: /api/v1
definitions:
...
    github_com_Angle-HR_server_internal_apidoc.ErrorBody:
        properties:
            code:
                example: VALIDATION_ERROR
                type: string
            details:
                additionalProperties: {}
                type: object
            message:
                example: invalid request body
                type: string
        type: object
    github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope:
        properties:
            error:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorBody'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    github_com_Angle-HR_server_internal_apidoc.Meta:
        properties:
            has_more:
                type: boolean
            next_cursor:
                type: string
            request_id:
                example: abc123
                type: string
        type: object
...
    internal_handler.AdminVerificationDetail:
        properties:
            address:
                $ref: '#/definitions/internal_handler.KYBAddress'
            address_type:
                type: string
            attempts:
                type: integer
            country_code:
                type: string
            identifiers:
                additionalProperties:
                    type: string
                type: object
            legal_name:
                type: string
            organization_id:
                type: string
            region:
                type: string
            registration_number:
                type: string
            status:
                type: string
            submitted_at:
                type: string
            tier:
                type: integer
        type: object
    internal_handler.AdminVerificationDetailEnvelope:
        properties:
            data:
                $ref: '#/definitions/internal_handler.AdminVerificationDetail'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    internal_handler.AdminVerificationItem:
        properties:
            country_code:
                type: string
            id:
                type: string
            organization_id:
                type: string
            region:
                type: string
            submitted_at:
                type: string
        type: object
    internal_handler.AdminVerificationListEnvelope:
        properties:
            data:
                items:
                    $ref: '#/definitions/internal_handler.AdminVerificationItem'
                type: array
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    internal_handler.AdminVerificationReviewData:
        properties:
            failure_reason:
                type: string
            organization_id:
                type: string
            status:
                enum:
                    - verified
                    - failed
                type: string
        type: object
    internal_handler.AdminVerificationReviewEnvelope:
        properties:
            data:
                $ref: '#/definitions/internal_handler.AdminVerificationReviewData'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    internal_handler.AdminVerificationReviewRequest:
        properties:
            decision:
                enum:
                    - approve
                    - reject
                type: string
            reason:
                enum:
                    - number_not_found
                    - name_mismatch
                    - address_mismatch
                    - inactive_entity
                    - wrong_country
                type: string
        type: object
...
    internal_handler.AuthLoginOTPRequest:
        properties:
            email:
                example: jerry@example.com
                type: string
        type: object
    internal_handler.AuthLoginOTPVerifyRequest:
        properties:
            code:
                example: "224879"
                type: string
            verification_session_id:
                example: 550e8400-e29b-41d4-a716-446655440000
                type: string
        type: object
...
    internal_handler.AuthSignupData:
        properties:
            code_expires_in_seconds:
                example: 300
                type: integer
            email:
                example: jerry@example.com
                type: string
            resend_available_in_seconds:
                example: 30
                type: integer
            verification_session_id:
                example: 550e8400-e29b-41d4-a716-446655440000
                type: string
        type: object
    internal_handler.AuthSignupEnvelope:
        properties:
            data:
                $ref: '#/definitions/internal_handler.AuthSignupData'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
...
    internal_handler.AuthTokenData:
        properties:
            access_token:
                example: eyJhbGciOiJIUzI1NiIs...
                type: string
            expires_in:
                example: 3600
                type: integer
            onboarding:
                $ref: '#/definitions/internal_handler.OnboardingProgressSummary'
            refresh_token:
                example: eyJhbGciOiJIUzI1NiIs...
                type: string
        type: object
    internal_handler.AuthTokenEnvelope:
        properties:
            data:
                $ref: '#/definitions/internal_handler.AuthTokenData'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
...
    internal_handler.KYBAddress:
        properties:
            city:
                type: string
            country:
                type: string
            line1:
                type: string
            line2:
                type: string
            post_code:
                type: string
            region:
                type: string
        type: object
...
    internal_handler.OnboardingProgressSummary:
        properties:
            completed_steps:
                example:
                    - verify_email
                    - profile
                items:
                    type: string
                type: array
            current_step:
                example: profile
                type: string
            next_step:
                example: compliance
                type: string
            status:
                enum:
                    - in_progress
                    - completed
                example: in_progress
                type: string
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: |-
        Waitlist and product onboarding API for Open HR.
        Interactive docs (Scalar) are served at `/` in non-production environments.
        Product flow: signup → verify email (6-digit OTP, 5 min expiry) → profile → address → [business] → complete.
        Onboarding step endpoints require `Authorization: Bearer <access_token>` after email verification.
        Admin ops console APIs live under `/api/v1/admin` and require a separate admin JWT (`token_use=admin`).
    title: Angle HR API
    version: "1.0"
paths:
...
    /admin/verification:
        get:
            parameters:
                - description: Page size (max 100)
                  in: query
                  name: limit
                  type: integer
                - description: Offset
                  in: query
                  name: offset
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminVerificationListEnvelope'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
            security:
                - BearerAuth: []
            summary: List pending company verification reviews
            tags:
                - admin/verification
    /admin/verification/{organizationID}:
        get:
            description: Returns the submitted details from the organization's region so the operator can compare them with the government portal.
            parameters:
                - description: Organization ID
                  in: path
                  name: organizationID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminVerificationDetailEnvelope'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
            security:
                - BearerAuth: []
            summary: Get a pending company verification review
            tags:
                - admin/verification
    /admin/verification/{organizationID}/review:
        post:
            consumes:
                - application/json
            description: Records the operator's decision. A rejection needs a valid failure reason and emails the owner.
            parameters:
                - description: Organization ID
                  in: path
                  name: organizationID
                  required: true
                  type: string
                - description: Decision
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/internal_handler.AdminVerificationReviewRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminVerificationReviewEnvelope'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
            security:
                - BearerAuth: []
            summary: Approve or reject a company verification
            tags:
                - admin/verification
...
    /auth/login/otp/request:
        post:
            consumes:
                - application/json
            description: Sends a 6-digit sign-in code. Separate from signup email verification.
            parameters:
                - description: Login OTP request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/internal_handler.AuthLoginOTPRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AuthSignupEnvelope'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "500":
                    description: Internal Server Error
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
            summary: Request passwordless login OTP
            tags:
                - auth
    /auth/login/otp/verify:
        post:
            consumes:
                - application/json
            description: Exchanges a login OTP for JWTs (or MFA challenge when TOTP is enabled).
            parameters:
                - description: Login OTP verify
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/internal_handler.AuthLoginOTPVerifyRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AuthTokenEnvelope'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
                "500":
                    description: Internal Server Error
                    schema:
                        $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.ErrorEnvelope'
            summary: Verify passwordless login OTP
            tags:
                - auth
...
securityDefinitions:
    BearerAuth:
        description: JWT access token. Send the raw token, or optionally prefix with "Bearer ".
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"
...